Moved client code to separate component for all guide pages in order to satisfy build error regarding suspense boundary with search params
