+++ app/guide/page.tsx
@@ -1,16 +1,14 @@
-"use client";
+import { Suspense } from "react";
+import { GuideContent } from "../components/guide/GuideContent";
 
-import { redirect } from 'next/navigation';
-import { useSearchParams } from "next/navigation";
-
-const Guide = () => {
-	const searchParams = useSearchParams();
-	const gameId = searchParams.get("gameId");
-	if (gameId) {
-		return redirect(`/guide/1?gameId=${gameId}`);
-	} else {
-		return redirect("/guide/1");
-	}
-};
-
-export default Guide;
+export default function Guide() {
+	return (
+		<Suspense fallback={
+			<div className="flex items-center justify-center h-full">
+				<div className="animate-spin rounded-full h-32 w-32 border-t-2 border-b-2 border-gray-900 dark:border-gray-100"></div>
+			</div>
+		}>
+			<GuideContent />
+		</Suspense>
+	);
+}
